portfolio-simulator-api/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database dependencies
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL driver

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data processing and analysis
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4

# HTTP client for external APIs
aiohttp==3.9.1
httpx==0.25.2

# Validation and configuration
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Report generation
reportlab==4.0.7
openpyxl==3.1.2

# Caching and background tasks
redis==5.0.1
celery==5.3.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Environment and utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# CORS handling
python-multipart==0.0.6 